+++ src/Application/Cli.php
@@ -158,7 +158,7 @@ final class Cli implements Implementation
      * @psalm-mutation-free
      */
     #[\Override]
-    public function route(string $pattern, callable $handle): self
+    public function route(callable $handle): self
     {
         return $this;
     }
@@ -167,7 +167,7 @@ final class Cli implements Implementation
      * @psalm-mutation-free
      */
     #[\Override]
-    public function appendRoutes(callable $append): self
+    public function mapRoute(callable $map): self
     {
         return $this;
     }
@@ -176,7 +176,7 @@ final class Cli implements Implementation
      * @psalm-mutation-free
      */
     #[\Override]
-    public function mapRequestHandler(callable $map): self
+    public function routeNotFound(callable $handle): self
     {
         return $this;
     }
@@ -185,7 +185,7 @@ final class Cli implements Implementation
      * @psalm-mutation-free
      */
     #[\Override]
-    public function notFoundRequestHandler(callable $handle): self
+    public function recoverRouteError(callable $recover): self
     {
         return $this;
     }