+++ src/Application/Implementation.php
@@ -3,11 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Framework\Application;
 
-use Innmind\Framework\{
-    Environment,
-    Http\Routes,
-    Http\RequestHandler,
-};
+use Innmind\Framework\Environment;
 use Innmind\OperatingSystem\OperatingSystem;
 use Innmind\CLI\{
     Environment as CliEnv,
@@ -17,7 +13,10 @@ use Innmind\DI\{
     Container,
     Service,
 };
-use Innmind\Router\Component;
+use Innmind\Router\{
+    Component,
+    Pipe,
+};
 use Innmind\Http\{
     ServerRequest,
     Response,
@@ -82,39 +81,38 @@ interface Implementation
     /**
      * @psalm-mutation-free
      *
-     * @param literal-string $pattern
-     * @param callable(Container, OperatingSystem, Environment): Component<SideEffect, Response> $handle
+     * @param callable(Pipe, Container, OperatingSystem, Environment): Component<SideEffect, Response> $handle
      *
      * @return self<I, O>
      */
-    public function route(string $pattern, callable $handle): self;
+    public function route(callable $handle): self;
 
     /**
      * @psalm-mutation-free
      *
-     * @param callable(Routes, Container, OperatingSystem, Environment): Routes $append
+     * @param callable(Component<SideEffect, Response>, Container): Component<SideEffect, Response> $map
      *
      * @return self<I, O>
      */
-    public function appendRoutes(callable $append): self;
+    public function mapRoute(callable $map): self;
 
     /**
      * @psalm-mutation-free
      *
-     * @param callable(RequestHandler, Container, OperatingSystem, Environment): RequestHandler $map
+     * @param callable(ServerRequest, Container, OperatingSystem, Environment): Attempt<Response> $handle
      *
      * @return self<I, O>
      */
-    public function mapRequestHandler(callable $map): self;
+    public function routeNotFound(callable $handle): self;
 
     /**
      * @psalm-mutation-free
      *
-     * @param callable(ServerRequest, Container, OperatingSystem, Environment): Response $handle
+     * @param callable(ServerRequest, \Throwable, Container): Attempt<Response> $recover
      *
      * @return self<I, O>
      */
-    public function notFoundRequestHandler(callable $handle): self;
+    public function recoverRouteError(callable $recover): self;
 
     /**
      * @param I $input