+++ CHANGELOG.md
@@ -2,16 +2,33 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Innmind\Framework\Http\Route`
+- `Innmind\Framework\Http\Route\Reference`
+- `Innmind\Framework\Application::mapRoute()`
+- `Innmind\Framework\Application::routes(class-string<Innmind\Framework\Http\Route\Reference>)`
+- `Innmind\Framework\Application::recoverRouteError()`
+
 ### Changed
 
 - Requires `innmind/foundation:~1.9`
 - Requires `innmind/di:~3.0`
 - `Innmind\Framework\Application::route()` callable must now return a `Innmind\Router\Component`
+- `Innmind\Framework\Application::route()` callable first parameter now is a `Innmint\Router\Pipe`
+- `Innmind\Framework\Application::route()` first parameter must now be expressed via a component inside the callable
+- `Innminf\Framework\Application::notFoundRequestHandler()` callable must now return an `Innmind\Immutable\Attempt<Response>`
+- `Innminf\Framework\Application::notFoundRequestHandler()` has been renamed `::routeNotFound()`
 
 ### Removed
 
 - The ability to use `string`s to reference services
 - `Innmind\Framework\Http\Service`
+- `Innmind\Framework\Http\To`
+- `Innmind\Framework\Http\Routes`
+- `Innmind\Framework\Application::appendRoutes()`
+- `Innmind\Framework\Application::mapRequestHandler()`
+- `Innmind\Framework\Http\RequestHandler`
 
 ### Fixed